Выносит createDestinationDescriptionTemplate из утилит в темлпейт
